requirements.txt
amqp==2.5.0
asn1crypto==0.24.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
billiard==3.6.0.0
cached-property==1.5.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
django-cors-headers==3.1.0
django-environ==0.4.5
djangorestframework==3.10.2
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
idna==2.7
jsonschema==2.6.0
kombu==4.6.3
paramiko==2.6.0
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.1
PyYAML==4.2b1
redis==3.3.6
requests==2.20.1
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
texttable==0.9.1
urllib3==1.24.3
vine==1.3.0
websocket-client==0.56.0
